hamburger button now collapses(for real this time)
this change was troublesome to implement even though bootstrap builds the navbar to make stuff like this easier. when the user clicks a nav-link or the hamburger button again, the collapsing nav menu will collapse as intended then bring you to the page you clicked on. looks much better for mobile now but there is still more work to do on mobile accessibility